+++ NEWS.md
@@ -2,7 +2,7 @@ interpretCI 0.1.2
 =================
 (2022-Feb-1)
 
-* new function draw_x2() added
+* new function draw_x2(), show_x2_table() added
 
 
 interpretCI 0.1.1